AddRestroom.test.tsx AddSimpleRestroom.test.tsxの作成
