+++ next/src/pages/index.test.tsx
@@ -1,16 +1,18 @@
-import { render, screen } from '@testing-library/react'
+import { render, screen, act ,rerender} from '@testing-library/react'
 import Index from '@/pages/index'
 import '@testing-library/jest-dom'
 
-//camelcase-keysのモックの作成
 jest.mock('camelcase-keys', () => ({
   default: jest.fn().mockReturnValue({}),
 }))
 
-describe('Index', () => {
-  it('Should display "map,"', () => {
-    render(<Index />)
 
-    expect(screen.getByTestId('map')).toBeTruthy()
+describe('Index', () => {
+  it('Should display "map"', async () => {
+    await act(async () => {
+      render(<Index />)
+    })
+
+    expect(screen.getByTestId('map')).toBeInTheDocument()
   })
 })
